Compile against latest Swiften
Compiles using the nested scons. For the .pro or Makefile
you'd need to remove the compilation of the BoostEventLoop.

Adds a filthy hack or two that can be backed out once
I've tested it runs

+++ LatencyWorkloadBenchmark.h
@@ -32,7 +32,7 @@ public:
 		int noOfActiveSessions;
 		int stanzasPerConnection;
 		int warmupStanzas;
-		int parallelLogins;
+		size_t parallelLogins;
 		std::string bodymessage;
 		bool noTLS;
 		bool noCompression;
